feat: Implement operator verification success and rejected modals, integrating them into the modal registry and operator profile.
