+++ src/features/modal/ui/OperatorVerification/OperatorVerificationRejected/constants.ts
@@ -1,0 +1,7 @@
+export const MODAL_VERIFICATION_REJECTED_TITLE = 'Вашу заявку відхилено!';
+export const MODAL_VERIFICATION_REJECTED_DESCRIPTION =
+  'Ваш статус “Туроператор” не підтверджено.';
+export const MODAL_VERIFICATION_REJECTED_REASON_LABEL = 'Причина';
+export const MODAL_VERIFICATION_REJECTED_BUTTON_TEXT = 'Подати нові дані';
+export const MODAL_VERIFICATION_REJECTED_PLACEHOLDER_TEXT =
+  'Тут буде вказана причина відхилення вашої заявки на статус "Туроператор"';
