+++ src/features/modal/ui/OperatorVerification/OperatorVerificationSuccess/constants.ts
@@ -1,0 +1,4 @@
+export const MODAL_VERIFICATION_SUCCESS_TITLE = 'Вітаємо!';
+export const MODAL_VERIFICATION_SUCCESS_DESCRIPTION =
+  'Ваш статус “Туроператор” підтверджено. Ви можете почати створювати свої тури';
+export const MODAL_VERIFICATION_SUCCESS_BUTTON_TEXT = 'Створити тур';
