+++ src/features/modal/ui/OperatorVerification/constants.ts
@@ -5,16 +5,3 @@ export const MODAL_VERIFICATION_FOOTER_BUTTON_TEXT =
 export const MODAL_VERIFICATION_PENDING_TITLE = 'Дякуємо!';
 export const MODAL_VERIFICATION_PENDING_DESCRIPTION =
   'Вашу заявку на статус “Туроператор” отримано. Очікуйте, будь ласка, підтвердження вашої веріфікації протягом доби';
-
-export const MODAL_VERIFICATION_SUCCESS_TITLE = 'Вітаємо!';
-export const MODAL_VERIFICATION_SUCCESS_DESCRIPTION =
-  'Ваш статус “Туроператор” підтверджено. Ви можете почати створювати свої тури';
-export const MODAL_VERIFICATION_SUCCESS_BUTTON_TEXT = 'Створити тур';
-
-export const MODAL_VERIFICATION_REJECTED_TITLE = 'Вашу заявку відхилено!';
-export const MODAL_VERIFICATION_REJECTED_DESCRIPTION =
-  'Ваш статус “Туроператор” не підтверджено.';
-export const MODAL_VERIFICATION_REJECTED_REASON_LABEL = 'Причина';
-export const MODAL_VERIFICATION_REJECTED_BUTTON_TEXT = 'Подати нові дані';
-export const MODAL_VERIFICATION_REJECTED_PLACEHOLDER_TEXT =
-  'Тут буде вказана причина відхилення вашої заявки на статус "Туроператор"';
