+++ src/pages-layer/operator-profile/ui/OperatorProfilePage.tsx
@@ -1,5 +1,11 @@
+'use client';
+
 import { OperatorProfile } from '@/features/operator-profile';
 
+import { useOperatorVerificationModal } from '../model/useOperatorVerificationModal';
+
 export const OperatorProfilePage = () => {
+  useOperatorVerificationModal();
+
   return <OperatorProfile />;
 };
