--- src/pages-layer/operator/index.ts
@@ -1,1 +1,0 @@
-export { OperatorPage } from './ui/OperatorPage';
